kousin

+++ mogupro/game/src/Game/Field/cChunk.cpp
@@ -60,7 +60,7 @@ cBlock* cChunk::getBlock( const int& x, const int& y, const int& z )
 cBlock* cChunk::getBlock( const ci::ivec3& c )
 {
     int y = c.y / CHUNK_SIZE;
-    if ( y >= mChunkLayers.size() || y < 0 )
+    if ( y >= (int) mChunkLayers.size() || y < 0 )
         return nullptr;
     auto& layer = mChunkLayers[y];
     y = c.y % CHUNK_SIZE;
